+++ plugins/tasks/http/src/main/java/com/walmartlabs/concord/plugins/http/SimpleHttpClient.java
@@ -149,6 +149,9 @@ public class SimpleHttpClient {
         }
 
         HttpEntity e = r.getEntity();
+        if (e == null) {
+            return null;
+        }
 
         switch (t) {
             case FILE: